Basic auto-configuration
This creates a very basic (dumb) auto-configuration. It kicks in
unconditionally and erases our own bean definition. This is certainly not
what we want to do. We'll improve the auto-configuration in the next
commit.

This showcases that an auto-configuration is a plain configuration that
is registered in `META-INF/spring.factories`.